backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
opencv-python>=4.7.0
Pillow>=9.5.0
librosa>=0.10.0
soundfile>=0.12.0
easyocr>=1.7.0
moviepy>=1.0.3
sentencepiece>=0.1.99
protobuf>=4.23.0
numpy>=1.24.0
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

